requirements.txt
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.4.0
matplotlib>=3.8.0
pyyaml>=6.0.1